+++ sdk/keyvault/azure-security-keyvault-secrets/test/ut/secret_get_client_deserialize_test.cpp
@@ -2,101 +2,14 @@
 // SPDX-License-Identifier: MIT
 
 #include "../src/private/secret_serializers.hpp"
-#include <azure/core/http/http.hpp>
-#include <azure/core/http/policies/policy.hpp>
-#include <azure/keyvault/secrets/keyvault_secret.hpp>
+#include "secret_get_client_deserialize_test.hpp"
+
 #include <azure/keyvault/secrets/secret_client.hpp>
-#include <gtest/gtest.h>
-#include <string>
 
 using namespace Azure::Security::KeyVault::Secrets;
 using namespace Azure::Security::KeyVault::Secrets::_detail;
-using namespace Azure::Core::Http::_internal;
 
-namespace {
-Azure::Core::Http::RawResponse getPartialResponse()
-{
-  auto response = Azure::Core::Http::RawResponse(1, 1, Azure::Core::Http::HttpStatusCode::Ok, "OK");
-
-  constexpr static const uint8_t responseBody[] = R"json({
-        "value": "mysecretvalue",
-        "id": "https://myvault.vault.azure.net/secrets/mysecretname/4387e9f3d6e14c459867679a90fd0f79",
-        "attributes": {
-          "enabled": true,
-          "created": 1493938410,
-          "updated": 1493938410,
-          "recoveryLevel": "Recoverable+Purgeable"
-         }
-    }
-)json";
-
-  response.SetHeader(HttpShared::ContentType, "application/json");
-  response.SetHeader(HttpShared::MsRequestId, "1");
-  response.SetHeader(HttpShared::MsClientRequestId, "2");
-  response.SetBody(std::vector<uint8_t>(responseBody, responseBody + sizeof(responseBody)));
-  response.SetBodyStream(
-      std::make_unique<Azure::Core::IO::MemoryBodyStream>(responseBody, sizeof(responseBody) - 1));
-
-  return response;
-}
-
-Azure::Core::Http::RawResponse getFullResponse()
-
-{
-  auto response = Azure::Core::Http::RawResponse(1, 1, Azure::Core::Http::HttpStatusCode::Ok, "OK");
-
-  constexpr static const uint8_t responseBody[] = R"json({
-        "value": "mysecretvalue",
-        "id": "https://myvault.vault.azure.net/secrets/mysecretname/4387e9f3d6e14c459867679a90fd0f79",
-        "contentType" : "ct",
-        "kid": "kid",
-        "managed": true,
-        "attributes": {
-          "enabled": true,
-          "created": 1493938410,
-          "updated": 1493938410,
-          "recoveryLevel": "Recoverable+Purgeable"
-         }
-    }
-)json";
-
-  response.SetHeader(HttpShared::ContentType, "application/json");
-  response.SetHeader(HttpShared::MsRequestId, "1");
-  response.SetHeader(HttpShared::MsClientRequestId, "2");
-  response.SetBody(std::vector<uint8_t>(responseBody, responseBody + sizeof(responseBody)));
-  response.SetBodyStream(
-      std::make_unique<Azure::Core::IO::MemoryBodyStream>(responseBody, sizeof(responseBody) - 1));
-
-  return response;
-}
-
-void runPartialExpect(KeyVaultSecret& secret) {
-  EXPECT_EQ(secret.Value, "mysecretvalue");
-  EXPECT_EQ(
-      secret.Id,
-      "https://myvault.vault.azure.net/secrets/mysecretname/4387e9f3d6e14c459867679a90fd0f79");
-  EXPECT_EQ(secret.KeyId.HasValue(), false);
-  EXPECT_EQ(secret.Properties.Enabled.Value(), true);
-  EXPECT_EQ(secret.Managed, false);
-  EXPECT_EQ(secret.Properties.UpdatedOn.HasValue(), true);
-  EXPECT_EQ(secret.Properties.CreatedOn.HasValue(), true);
-}
-
-void runFullExpect(KeyVaultSecret& secret) {
-  EXPECT_EQ(secret.Value, "mysecretvalue");
-  EXPECT_EQ(
-      secret.Id,
-      "https://myvault.vault.azure.net/secrets/mysecretname/4387e9f3d6e14c459867679a90fd0f79");
-  EXPECT_EQ(secret.Properties.Enabled.Value(), true);
-  EXPECT_EQ(secret.Managed, true);
-  EXPECT_EQ(secret.ContentType.Value(), "ct");
-  EXPECT_EQ(secret.KeyId.Value(), "kid");
-  EXPECT_EQ(secret.Properties.UpdatedOn.HasValue(), true);
-  EXPECT_EQ(secret.Properties.CreatedOn.HasValue(), true);
-}
-
-} // namespace
-TEST(SecretClient, GetClientDeserializePartial1)
+TEST(KeyVaultSecretSerializer, GetClientDeserializePartial1)
 {
   auto response = getPartialResponse();
 
@@ -105,7 +18,7 @@ TEST(SecretClient, GetClientDeserializePartial1)
   runPartialExpect(secret);
 }
 
-TEST(SecretClient, GetClientDeserializePartial2)
+TEST(KeyVaultSecretSerializer, GetClientDeserializePartial2)
 {
   auto response = getPartialResponse();
 
@@ -116,7 +29,7 @@ TEST(SecretClient, GetClientDeserializePartial2)
   runPartialExpect(secret);
 }
 
-TEST(SecretClient, GetClientDeserializePartial3)
+TEST(KeyVaultSecretSerializer, GetClientDeserializePartial3)
 {
   auto response = getPartialResponse();
 
@@ -127,7 +40,7 @@ TEST(SecretClient, GetClientDeserializePartial3)
   runPartialExpect(secret);
 }
 
-TEST(SecretClient, GetClientdeserializeFull1)
+TEST(KeyVaultSecretSerializer, GetClientdeserializeFull1)
 {
   auto response = getFullResponse();
 
@@ -136,7 +49,7 @@ TEST(SecretClient, GetClientdeserializeFull1)
   runFullExpect(secret);
 }
 
-TEST(SecretClient, GetClientdeserializeFull2)
+TEST(KeyVaultSecretSerializer, GetClientdeserializeFull2)
 {
   auto response = getFullResponse();
 
@@ -147,7 +60,7 @@ TEST(SecretClient, GetClientdeserializeFull2)
   runFullExpect(secret);
 }
 
-TEST(SecretClient, GetClientdeserializeFull3)
+TEST(KeyVaultSecretSerializer, GetClientdeserializeFull3)
 {
   auto response = getFullResponse();
